Fixed minor bugs
To fix minor bugs, you must generate bugs.  To generate bugs, you must be bugged yourself before you can even be a minor; WHAT.  Not to fix a minor bug, but to be bugged to fix what was bugged in the first place.  Now let's question, to debug the minor bug.